services/vuln-detector/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pyyaml==6.0.1
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
pydantic==2.5.2
